streaming-llm-anomaly-detector/requirements.txt
# Requirements for Streaming LLM Anomaly Detector and Incident Q&A Chatbot

# Core dependencies (Updated to working versions)
openai==2.15.0
python-dotenv==1.0.1

# LangChain ecosystem for chatbot
langchain==0.1.20
langchain-openai==1.1.7
langchain-core==1.2.7
langchain-community==0.0.38
langchain-chroma==0.1.4
langgraph==1.0.5

# Vector database and document processing
chromadb
pypdf

# Data processing
pandas

# Database
psycopg2-binary

# Web Interface
streamlit
fastapi
uvicorn
plotly